% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data for all districts}
\usage{
get_data(year = "2021", month = "10")
}
\arguments{
\item{year}{A string or numeric value containing the year.}

\item{month}{A string or numeric value containing the month.}
}
\value{
A data.frame.
}
\description{
Get data for all districts
}
\examples{
get_data(id = c("09571", "05558"), year = "2021", month = "10")
}
