% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_district.R
\name{get_district}
\alias{get_district}
\title{Get labour data for district and timestamp}
\usage{
get_district(id = "09571", year = "2021", month = "10")
}
\arguments{
\item{id}{A string or numeric value containing the district id.}

\item{year}{A string or numeric value containing the year.}

\item{month}{A string or numeric value containing the month.}
}
\value{
A data.frame.
}
\description{
Get labour data for district and timestamp
}
\examples{
get_district(id = "09571", year = "2021", month = "10")
}
